On the fly fix up - app engine needs a main module.
